Create NotificationFactory
