+++ libraries/push/impl/src/main/kotlin/io/element/android/libraries/push/impl/notifications/NotificationFactory.kt
@@ -19,6 +19,7 @@ package io.element.android.libraries.push.impl.notifications
 import android.app.Notification
 import io.element.android.libraries.matrix.api.core.RoomId
 import io.element.android.libraries.matrix.api.core.SessionId
+import io.element.android.libraries.push.impl.notifications.factories.NotificationFactory
 import io.element.android.libraries.push.impl.notifications.model.InviteNotifiableEvent
 import io.element.android.libraries.push.impl.notifications.model.NotifiableMessageEvent
 import io.element.android.libraries.push.impl.notifications.model.SimpleNotifiableEvent
@@ -26,8 +27,9 @@ import javax.inject.Inject
 
 private typealias ProcessedMessageEvents = List<ProcessedEvent<NotifiableMessageEvent>>
 
+// TODO Find a better name, it clashes with io.element.android.libraries.push.impl.notifications.factories.NotificationFactory
 class NotificationFactory @Inject constructor(
-    private val notificationUtils: NotificationUtils,
+    private val notificationFactory: NotificationFactory,
     private val roomGroupMessageCreator: RoomGroupMessageCreator,
     private val summaryGroupMessageCreator: SummaryGroupMessageCreator
 ) {
@@ -66,7 +68,7 @@ class NotificationFactory @Inject constructor(
             when (processed) {
                 ProcessedEvent.Type.REMOVE -> OneShotNotification.Removed(key = event.roomId.value)
                 ProcessedEvent.Type.KEEP -> OneShotNotification.Append(
-                    notificationUtils.buildRoomInvitationNotification(event),
+                    notificationFactory.createRoomInvitationNotification(event),
                     OneShotNotification.Append.Meta(
                         key = event.roomId.value,
                         summaryLine = event.description,
@@ -84,7 +86,7 @@ class NotificationFactory @Inject constructor(
             when (processed) {
                 ProcessedEvent.Type.REMOVE -> OneShotNotification.Removed(key = event.eventId.value)
                 ProcessedEvent.Type.KEEP -> OneShotNotification.Append(
-                    notificationUtils.buildSimpleEventNotification(event),
+                    notificationFactory.createSimpleEventNotification(event),
                     OneShotNotification.Append.Meta(
                         key = event.eventId.value,
                         summaryLine = event.description,
